+++ src/main/java/com/api/customer/respositories/TechnicianRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.api.customer.entities.Technician;
 
-public interface TechnicianRepository extends JpaRepository<Technician, Integer> {
+public interface TechnicianRepository extends JpaRepository<Technician, Long> {
 
 }
